Fix login into mantis ending with access denied.
Doku wiki version 2024-02-06b interacts with mantisbt-2.25.7 incorrectly using the access_get_project_level function.

Dokuwiki uses this API internallly for users other than the current logged in user and not specifying the ID breaks that functionality.

Similarly this also does not work for the session when a user is being authenticated as some global value is not actually set in the mantis project and thus the function fails as no user is found.